+++ lib/settings/metaboxes/sections.php
@@ -114,9 +114,11 @@ function mai_do_sections_metabox() {
 		'default' => 'md',
 		'options' => array(
 			'auto' => __( 'Auto (Use height of content)', 'mai-pro-engine' ),
+			'xs'   => __( 'Extra Small', 'mai-pro-engine' ),
 			'sm'   => __( 'Small', 'mai-pro-engine' ),
-			'md'   => __( 'Medium', 'mai-pro-engine' ),
+			'md'   => __( 'Medium (Default)', 'mai-pro-engine' ),
 			'lg'   => __( 'Large', 'mai-pro-engine' ),
+			'xl'   => __( 'Extra Large', 'mai-pro-engine' ),
 		),
 	) );