+++ mai-pro-engine.php
@@ -5,7 +5,7 @@
  * Plugin URI:      https://maipro.io/
  * Description:     The Mai Pro Engine plugin
  *
- * Version:         1.1.6
+ * Version:         1.1.6.1
  *
  * GitHub URI:      maiprowp/mai-pro-engine
  *
@@ -91,10 +91,10 @@ final class Mai_Pro_Engine {
 	private function setup_constants() {
 
 		// Plugin version.
-		define( 'MAI_PRO_ENGINE_VERSION', '1.1.6' );
+		define( 'MAI_PRO_ENGINE_VERSION', '1.1.6.1' );
 
 		// DB version.
-		define( 'MAI_PRO_ENGINE_DB_VERSION', '1160' );
+		define( 'MAI_PRO_ENGINE_DB_VERSION', '1161' );
 
 		// Plugin Folder Path.
 		define( 'MAI_PRO_ENGINE_PLUGIN_DIR', plugin_dir_path( __FILE__ ) );